requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
fair-esm>=2.0.0

# GFlowNet
torchgfn>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
biopython>=1.81

# Evaluation and visualization
scikit-learn>=1.2.0
umap-learn>=0.5.0
hdbscan>=0.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Experiment tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Development
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0

# Notebooks (optional)
jupyter>=1.0.0
ipywidgets>=8.0.0
